Unfinished refactoring and cleanup of provider settings
