+++ app/components/settings/SettingsWindow.tsx
@@ -12,6 +12,7 @@ import commit from '~/commit.json';
 import Cookies from 'js-cookie';
 import styles from './Settings.module.scss';
 import { Switch } from '~/components/ui/Switch';
+import { getProvidersInitialState } from '~/components/settings/getProvidersInitialState';
 
 interface SettingsProps {
   open: boolean;
@@ -81,25 +82,7 @@ export const SettingsWindow = ({ open, onClose }: SettingsProps) => {
   ];
 
   // Load providers from cookies on mount
-  const [providers, setProviders] = useState(() => {
-    const savedProviders = Cookies.get('providers');
-
-    if (savedProviders) {
-      try {
-        const parsedProviders = JSON.parse(savedProviders);
-
-        // Merge saved enabled states with the base provider list
-        return providersList.map((provider) => ({
-          ...provider,
-          isEnabled: parsedProviders[provider.name] || false,
-        }));
-      } catch (error) {
-        console.error('Failed to parse providers from cookies:', error);
-      }
-    }
-
-    return providersList;
-  });
+  const [providers, setProviders] = useState(getProvidersInitialState(providersList));
 
   const handleToggleProvider = (providerName: string, enabled: boolean) => {
     setProviders((prevProviders) => {
@@ -372,16 +355,14 @@ export const SettingsWindow = ({ open, onClose }: SettingsProps) => {
                         <h3 className="text-lg font-medium text-bolt-elements-textPrimary mb-4">Optional Features</h3>
                         <div className="flex items-center justify-between mb-2">
                           <span className="text-bolt-elements-textPrimary">Debug Info</span>
-                          <Switch
-                            className="ml-auto"
-                            checked={isDebugEnabled}
-                            onCheckedChange={handleToggleDebug}
-                          />
+                          <Switch className="ml-auto" checked={isDebugEnabled} onCheckedChange={handleToggleDebug} />
                         </div>
                       </div>
 
                       <div className="mb-6 border-t border-bolt-elements-borderColor pt-4">
-                        <h3 className="text-lg font-medium text-bolt-elements-textPrimary mb-4">Experimental Features</h3>
+                        <h3 className="text-lg font-medium text-bolt-elements-textPrimary mb-4">
+                          Experimental Features
+                        </h3>
                         <p className="text-sm text-bolt-elements-textSecondary mb-4">
                           Disclaimer: Experimental features may be unstable and are subject to change.
                         </p>
@@ -439,7 +420,9 @@ export const SettingsWindow = ({ open, onClose }: SettingsProps) => {
                       <h3 className="text-lg font-medium text-bolt-elements-textPrimary mb-4">GitHub Connection</h3>
                       <div className="flex mb-4">
                         <div className="flex-1 mr-2">
-                          <label className="block text-sm text-bolt-elements-textSecondary mb-1">GitHub Username:</label>
+                          <label className="block text-sm text-bolt-elements-textSecondary mb-1">
+                            GitHub Username:
+                          </label>
                           <input
                             type="text"
                             value={githubUsername}
@@ -448,7 +431,9 @@ export const SettingsWindow = ({ open, onClose }: SettingsProps) => {
                           />
                         </div>
                         <div className="flex-1">
-                          <label className="block text-sm text-bolt-elements-textSecondary mb-1">Personal Access Token:</label>
+                          <label className="block text-sm text-bolt-elements-textSecondary mb-1">
+                            Personal Access Token:
+                          </label>
                           <input
                             type="password"
                             value={githubToken}
